draft: Start Trying to Create a Way to Upload Sermons (INCOMPLETE)
We've made a lot of headway trying to provide a means to upload
sermons. However, we're running into a bug with the PodBean API
or a really obscure syntax error on our end. And it isn't clear.
We're in communication with the PodBean help team at the moment.
Committing current progress so it doesn't get lost.

Current Progress:
* Create a route to validate the form data, update file info,
  and attempt to publish a new episode.
* Create a means to upload files via `/files/uploadAuthorize`
  as mentioned in the docs.

Remaining:
* Need `podbeanService.upload` to successfully generate
  new episodes. We're getting an obscure error about
  the episode type right now, even though the type should
  be correct.